Day6 completed with tests Part2 very ugly.. bleeee..

+++ src/main/java/pl/pacinho/adventofcode2021/challange/CalculateI.java
@@ -1,5 +1,5 @@
 package pl.pacinho.adventofcode2021.challange;
 
 public interface CalculateI {
-       int calculate(String filePath);
+       long calculate(String filePath);
 }
