+++ src/main/java/pl/pacinho/adventofcode2021/challange/day5/Day5Part1.java
@@ -15,7 +15,7 @@ public class Day5Part1 implements CalculateI {
     }
 
     @Override
-    public int calculate(String filePath) {
+    public long calculate(String filePath) {
         List<String> lines = FileUtils.readTxt(new File(filePath));
         List<String> points = new ArrayList<>();
